made changes for interview task.
