+++ .gitignore
@@ -1,3 +1,5 @@
 build
 .idea
 .gradle
+/target/
+/bin/
