+++ src/main/java/com/codinground/test/SignInTest.java
@@ -1,0 +1,38 @@
+package com.codinground.test;
+import org.openqa.selenium.By;
+import org.testng.Assert;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+
+import com.codinground.utils.SeleniumUtils;
+
+public class SignInTest extends SeleniumUtils{
+
+	@BeforeClass
+	public void init() {
+		loadBrowser();
+	}
+
+    @Test
+    public void shouldThrowAnErrorIfSignInDetailsAreMissing() {
+
+        driver.get("https://www.cleartrip.com/");
+        waitFor(2000);
+
+        driver.findElement(By.linkText("Your trips")).click();
+        driver.findElement(By.id("SignIn")).click();
+        driver.switchTo().frame("modal_window");
+        driver.findElement(By.id("signInButton")).click();
+
+        String errors1 = driver.findElement(By.id("errors1")).getText();
+        Assert.assertTrue(errors1.contains("There were errors in your submission"));
+    }
+
+    @AfterClass
+    public void afterClass() {
+    	tearDown();
+    }
+   
+
+}
