+++ testng.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="Suite">
+  <test thread-count="5" name="Test">
+    <classes>
+      <class name="com.codinground.test.FlightBookingTest"/>
+      <class name="com.codinground.utils.SeleniumUtils"/>
+      <class name="com.codinground.test.HotelBookingTest"/>
+      <class name="com.codinground.test.SignInTest"/>
+    </classes>
+  </test> <!-- Test -->
+</suite> <!-- Suite -->
